+++ src/RhinoInside.Revit.External/DB/Extensions/Family.cs
@@ -8,5 +8,21 @@ namespace RhinoInside.Revit.External.DB.Extensions
     {
       return family.get_Parameter(BuiltInParameter.FAMILY_HOSTING_BEHAVIOR).AsEnum<FamilyHostingBehavior>();
     }
+
+    public static bool IsWorkPlaneBased(this Family family)
+    {
+      return family.get_Parameter(BuiltInParameter.FAMILY_WORK_PLANE_BASED)?.AsBoolean() ?? false;
+    }
+
+    public static bool SetWorkPlaneBased(this Family family, bool value)
+    {
+      using (var workPlaneBasedParameter = family.get_Parameter(BuiltInParameter.FAMILY_WORK_PLANE_BASED))
+      {
+        if (workPlaneBasedParameter?.IsReadOnly is false)
+          return workPlaneBasedParameter.Update(value);
+      }
+
+      return false;
+    }
   }
 }