+++ src/RhinoInside.Revit.GH/Components/Element/Components/AddComponent-Adaptive.cs
@@ -12,22 +12,22 @@ namespace RhinoInside.Revit.GH.Components
   using Kernel.Attributes;
   using External.DB.Extensions;
 
-  public class AdaptiveComponentByPoints : ReconstructElementComponent
+  public class AddComponentAdaptive : ReconstructElementComponent
   {
     public override Guid ComponentGuid => new Guid("E8DDC0E4-97E9-4659-9945-E8C77114273D");
     public override GH_Exposure Exposure => GH_Exposure.primary;
 
-    public AdaptiveComponentByPoints() : base
+    public AddComponentAdaptive() : base
     (
       name: "Add Component (Adaptive)",
-      nickname: "CompAdap",
+      nickname: "A-Component",
       description: "Given a collection of Points, it adds an AdaptiveComponent element to the active Revit document",
       category: "Revit",
       subCategory: "Build"
     )
     { }
 
-    void ReconstructAdaptiveComponentByPoints
+    void ReconstructAddComponentAdaptive
     (
       [Optional, NickName("DOC")]
       ARDB.Document document,